Add limited support for Google App Engine.
This allows the package to build and operate in a limited fashion
on Google App Engine, where "unsafe" package is not available.

Closes #35.
